Create README.md
Indian gemstones and jewellery enterprises are predominantly wedding ceremony driven. But because of the coronavirus, such occasions were cancelled, and wedding ceremony purchasing has come to a halt regardless of the continuing wedding ceremony season. Still, the rate of metals like gold have risen all of a sudden, and because of this, the call for a gold mortgage has instead expanded. India's common long-time period gold demand is likely around 850 tonnes due to its affinity to gold because of financial and social requirements.